move input classes to runner modules so that input definitions are close to code that uses them
